fix: auth_managerを削除して、ViewModel内に記載しました
